feat: Implement splash screen and update UI components for login
