+++ app/src/main/java/com/example/sora/friends/FriendScreen.kt
@@ -1,17 +1,12 @@
 package com.example.sora.friends
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -20,22 +15,18 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import coil.compose.AsyncImage
-
 
 // Mutual friend data class
-private data class MutualFriend(val name: String, val profileImageUrl: String)
+data class MutualFriend(val name: String, val profileImageUrl: String)
 
 // Dummy user data class
-private data class User(
+data class User(
     val id: Int,
     val name: String,
     val handle: String,
@@ -128,80 +119,6 @@ fun FriendScreen(navController: NavController? = null) {
                 })
             }
         }
-    }
-}
-
-@Composable
-private fun UserRow(user: User, onFollowClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .padding(12.dp)
-            .fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        AsyncImage(
-            model = user.profileImageUrl,
-            contentDescription = "Profile picture of ${user.name}",
-            modifier = Modifier
-                .background(Color.LightGray, RoundedCornerShape(50))
-                .height(48.dp)
-                .width(48.dp)
-        )
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = user.name,
-                fontWeight = FontWeight.W600,
-                fontSize = 15.sp
-            )
-            Text(
-                text = user.handle,
-                fontWeight = FontWeight.W400,
-                fontSize = 13.sp,
-                color = Color.Gray
-            )
-            if (user.mutualFriends.isNotEmpty()) {
-                MutualFriendsRow(user.mutualFriends)
-            }
-        }
-        Button(
-            onClick = onFollowClick,
-            shape = RoundedCornerShape(20.dp),
-            enabled = !user.isFollowed
-        ) {
-            Text(if (user.isFollowed) "Following" else "Follow")
-        }
-    }
-}
-
-@Composable
-private fun MutualFriendsRow(mutualFriends: List<MutualFriend>) {
-    Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.padding(top = 4.dp)) {
-        // Show up to 3 avatars, side by side (no overlap for compatibility)
-        mutualFriends.take(3).forEach { friend ->
-            AsyncImage(
-                model = friend.profileImageUrl,
-                contentDescription = "Mutual friend: ${friend.name}",
-                modifier = Modifier
-                    .height(20.dp)
-                    .width(20.dp)
-                    .background(Color.LightGray, RoundedCornerShape(50))
-                    .padding(end = 4.dp)
-            )
-        }
-        val names = mutualFriends.map { it.name }
-        val summary = when (names.size) {
-            1 -> "Followed by ${names[0]}"
-            2 -> "Followed by ${names[0]} and ${names[1]}"
-            3 -> "Followed by ${names[0]}, ${names[1]} and 1 other"
-            else -> "Followed by ${names[0]}, ${names[1]} and ${names.size - 2} others"
-        }
-        Text(
-            text = summary,
-            fontSize = 11.sp,
-            color = Color.Gray,
-            modifier = Modifier.padding(start = 8.dp)
-        )
     }
 }
